+++ tests/integration/mylib/include/mylib/mylib.h
@@ -1,0 +1,7 @@
+#ifndef MYLIB_MYLIB_H
+#define MYLIB_MYLIB_H
+
+int mylib_add(int a, int b);
+int mylib_subtract(int a, int b);
+
+#endif /* MYLIB_MYLIB_H */